Initial pass at cleaning up paywall config
more simplifications on the configuration for build scripts
