+++ paywall/next.config.js
@@ -4,42 +4,9 @@ const { join } = require('path')
 const { promisify } = require('util')
 const withCSS = require('@zeit/next-css')
 const withTypescript = require('@zeit/next-typescript')
-const dotenv = require('dotenv')
-const path = require('path')
-
-const unlockEnv = process.env.UNLOCK_ENV || 'dev'
-
-dotenv.config({
-  path: path.resolve(__dirname, '..', `.env.${unlockEnv}.local`),
-})
+const configVariables = require('./environment')
 
 const copyFile = promisify(fs.copyFile)
-
-const requiredConfigVariables = {
-  unlockEnv,
-  readOnlyProvider: process.env.READ_ONLY_PROVIDER,
-  locksmithUri: process.env.LOCKSMITH_URI,
-  erc20ContractSymbol: process.env.ERC20_CONTRACT_SYMBOL,
-  erc20ContractAddress: process.env.ERC20_CONTRACT_ADDRESS,
-}
-
-const optionalConfigVariables = {
-  httpProvider: process.env.HTTP_PROVIDER,
-  debugMode: process.env.DEBUG,
-}
-
-Object.keys(requiredConfigVariables).forEach(configVariableName => {
-  if (!requiredConfigVariables[configVariableName]) {
-    if (['dev', 'test'].indexOf(requiredConfigVariables.unlockEnv) > -1) {
-      return console.error(
-        `The configuration variable ${configVariableName} is falsy.`
-      )
-    }
-    throw new Error(
-      `The configuration variable ${configVariableName} is falsy.`
-    )
-  }
-})
 
 // TODO: remove this when next gets their act together
 // see https://github.com/zeit/next-plugins/issues/392#issuecomment-475845330
@@ -61,8 +28,7 @@ function HACK_removeMinimizeOptionFromCssLoaders(config) {
 module.exports = withTypescript(
   withCSS({
     publicRuntimeConfig: {
-      ...optionalConfigVariables,
-      ...requiredConfigVariables,
+      ...configVariables,
     },
     webpack: config => {
       // Fixes npm packages that depend on `fs` module
