+++ paywall/unlock.1.0.js.webpack.config.js
@@ -1,33 +1,7 @@
 /* eslint no-console: 0 */
-const dotenv = require('dotenv')
 var path = require('path')
 const webpack = require('webpack')
-
-const unlockEnv = process.env.UNLOCK_ENV || 'dev'
-const debug = process.env.DEBUG ? 1 : 0
-
-dotenv.config({
-  path: path.resolve(__dirname, '..', `.env.${unlockEnv}.local`),
-})
-
-const requiredConfigVariables = {
-  unlockEnv,
-  paywallUrl: process.env.PAYWALL_URL,
-  usersIframeUrl: process.env.USER_IFRAME_URL,
-}
-
-Object.keys(requiredConfigVariables).forEach(configVariableName => {
-  if (!requiredConfigVariables[configVariableName]) {
-    if (['dev', 'test'].indexOf(requiredConfigVariables.unlockEnv) > -1) {
-      return console.error(
-        `The configuration variable ${configVariableName} is falsy.`
-      )
-    }
-    throw new Error(
-      `The configuration variable ${configVariableName} is falsy.`
-    )
-  }
-})
+const configVariables = require('./environment')
 
 module.exports = () => {
   return {
@@ -57,12 +31,7 @@ module.exports = () => {
     },
     plugins: [
       new webpack.DefinePlugin({
-        'process.env.UNLOCK_ENV': "'" + unlockEnv + "'",
-        'process.env.DEBUG': debug,
-        'process.env.PAYWALL_URL':
-          "'" + requiredConfigVariables.paywallUrl + "'",
-        'process.env.USER_IFRAME_URL':
-          "'" + requiredConfigVariables.usersIframeUrl + "'",
+        __ENVIRONMENT_VARIABLES__: JSON.stringify(configVariables),
       }),
     ],
   }
